:boom: Change installments object schema
