+++ Paymongo.Sharp/Features/PaymentMethods/Contracts/PaymentMethodOption.cs
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using Paymongo.Sharp.Features.CardInstallments.Entities;
+using Paymongo.Sharp.Features.CardInstallments.Contracts;
 
 namespace Paymongo.Sharp.Features.PaymentMethods.Entities
 {
