+++ Paymongo.Sharp/Interfaces/IPaymongoClient.cs
@@ -20,9 +20,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+using Paymongo.Sharp.Features.CardInstallments;
 using Paymongo.Sharp.Features.Checkouts;
 using Paymongo.Sharp.Features.Customers;
-using Paymongo.Sharp.Features.Installments;
 using Paymongo.Sharp.Features.Links;
 using Paymongo.Sharp.Features.PaymentIntents;
 using Paymongo.Sharp.Features.PaymentMethods;
