+++ Paymongo.Sharp/Paymongo.Sharp.csproj
@@ -51,4 +51,8 @@ OTHER CHANGES (Package Rework):
       <Resource Include="icon.png" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Features\Installments\Entities\" />
+    </ItemGroup>
+
 </Project>
